Adding Ability for TSMarkdownParser to Have Custom Paragraph Formatting Block
This normally wouldn't be necessary, but I would like to give the NSAttributedString that is generated to have the ability to set, for instance, text color of the entire string. It appears that if you parse a string for use in a UITextView and you want the paragraph text to have a different color than, for instance, a bold block, there is no way to do this (changing the text color of UITextView overrides any custom colors set in the NSAttributedString.